requirements.txt
google-genai
google-cloud-aiplatform
fastapi
uvicorn
websockets
pydantic
python-dotenv
numpy
pandas
plotly
scikit-learn
python-multipart
